complete move away from horrid id3lib, and to new hotness taglib
git-svn-id: file:///srv/svn/spinny@420 9787d06c-1a0a-0410-a88a-902e1a4f81df

+++ libs/ews/connection.cpp
@@ -20,7 +20,7 @@ namespace ews {
 		  reply_(this),
 		  socket_detached_(false)
 	{
-		BOOST_LOGL(www,debug) << "NEW CONNECTION: " << (int)this << std::endl;
+		BOOST_LOGL(www,debug) << "NEW CONNECTION: " << this << std::endl;
 	}
 
 	boost::shared_ptr<asio::ip::tcp::socket>
